Refactor Bond integration to remove duplication in Entity code and unit tests (PR feedback)
