+++ homeassistant/components/bond/entity.py
