+++ homeassistant/components/bond/fan.py
@@ -15,8 +15,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import Entity
 
-from .bond_entity import BondEntity
 from .const import DOMAIN
+from .entity import BondEntity
 from .utils import BondDevice, get_bond_devices
